+++ lib/src/config/setting_config/setting_config.g.dart
@@ -48,8 +48,6 @@ abstract class _$SettingConfigCWProxy {
 
   SettingConfig expandRatiosOfSameMinmax(List<double> expandRatiosOfSameMinmax);
 
-  SettingConfig hideIndicatorsInTimeChart(bool hideIndicatorsInTimeChart);
-
   /// This function **does support** nullification of nullable fields. All `null` values passed to `non-nullable` fields will be ignored. You can also use `SettingConfig(...).copyWith.fieldName(...)` to override fields one at a time with nullification support.
   ///
   /// Usage
@@ -76,7 +74,6 @@ abstract class _$SettingConfigCWProxy {
     bool autoStartLastPriceCountDownTimer,
     bool autoLoadMoreData,
     List<double> expandRatiosOfSameMinmax,
-    bool hideIndicatorsInTimeChart,
   });
 }
 
@@ -166,10 +163,6 @@ class _$SettingConfigCWProxyImpl implements _$SettingConfigCWProxy {
       this(expandRatiosOfSameMinmax: expandRatiosOfSameMinmax);
 
   @override
-  SettingConfig hideIndicatorsInTimeChart(bool hideIndicatorsInTimeChart) =>
-      this(hideIndicatorsInTimeChart: hideIndicatorsInTimeChart);
-
-  @override
 
   /// This function **does support** nullification of nullable fields. All `null` values passed to `non-nullable` fields will be ignored. You can also use `SettingConfig(...).copyWith.fieldName(...)` to override fields one at a time with nullification support.
   ///
@@ -198,7 +191,6 @@ class _$SettingConfigCWProxyImpl implements _$SettingConfigCWProxy {
     Object? autoStartLastPriceCountDownTimer = const $CopyWithPlaceholder(),
     Object? autoLoadMoreData = const $CopyWithPlaceholder(),
     Object? expandRatiosOfSameMinmax = const $CopyWithPlaceholder(),
-    Object? hideIndicatorsInTimeChart = const $CopyWithPlaceholder(),
   }) {
     return SettingConfig(
       opacity: opacity == const $CopyWithPlaceholder()
@@ -283,11 +275,6 @@ class _$SettingConfigCWProxyImpl implements _$SettingConfigCWProxy {
               ? _value.expandRatiosOfSameMinmax
               // ignore: cast_nullable_to_non_nullable
               : expandRatiosOfSameMinmax as List<double>,
-      hideIndicatorsInTimeChart:
-          hideIndicatorsInTimeChart == const $CopyWithPlaceholder()
-              ? _value.hideIndicatorsInTimeChart
-              // ignore: cast_nullable_to_non_nullable
-              : hideIndicatorsInTimeChart as bool,
     );
   }
 }
@@ -334,8 +321,6 @@ SettingConfig _$SettingConfigFromJson(Map<String, dynamic> json) =>
                   ?.map((e) => (e as num).toDouble())
                   .toList() ??
               const [0.1, 0.05],
-      hideIndicatorsInTimeChart:
-          json['hideIndicatorsInTimeChart'] as bool? ?? false,
     );
 
 Map<String, dynamic> _$SettingConfigToJson(SettingConfig instance) =>
@@ -362,5 +347,4 @@ Map<String, dynamic> _$SettingConfigToJson(SettingConfig instance) =>
           instance.autoStartLastPriceCountDownTimer,
       'autoLoadMoreData': instance.autoLoadMoreData,
       'expandRatiosOfSameMinmax': instance.expandRatiosOfSameMinmax,
-      'hideIndicatorsInTimeChart': instance.hideIndicatorsInTimeChart,
     };